Make task fitler public
